Unify sudo prefixing functions
Unify __prepend_sudo and __last_sudo into one function and shortcut.
When the commandline is empty just entering sudo isn't all that helpful,
thus when can use this case to instead run the functionality __last_sudo
provided, so fetching the last history line and prepending sudo (or
removing it if present).